+++ mlir/include/mlir/Bindings/Python/NanobindAdaptors.h
@@ -19,7 +19,6 @@
 #ifndef MLIR_BINDINGS_PYTHON_NANOBINDADAPTORS_H
 #define MLIR_BINDINGS_PYTHON_NANOBINDADAPTORS_H
 
-#include <atomic>
 #include <cstdint>
 #include <memory>
 #include <optional>
@@ -35,41 +34,6 @@
 namespace mlir {
 namespace python {
 namespace {
-
-// Safely calls Python initialization code on first use, avoiding deadlocks.
-template <typename T>
-class SafeInit {
-public:
-  typedef std::unique_ptr<T> (*F)();
-
-  explicit SafeInit(F init_fn) : initFn(init_fn) {}
-
-  T &get() {
-    if (T *result = output.load()) {
-      return *result;
-    }
-
-    // Note: init_fn() may be called multiple times if, for example, the GIL is
-    // released during its execution. The intended use case is for module
-    // imports which are safe to perform multiple times. We are careful not to
-    // hold a lock across init_fn() to avoid lock ordering problems.
-    std::unique_ptr<T> m = initFn();
-    {
-      nanobind::ft_lock_guard lock(mu);
-      if (T *result = output.load()) {
-        return *result;
-      }
-      T *p = m.release();
-      output.store(p);
-      return *p;
-    }
-  }
-
-private:
-  nanobind::ft_mutex mu;
-  std::atomic<T *> output{nullptr};
-  F initFn;
-};
 
 nanobind::module_ &irModule() {
   static SafeInit<nanobind::module_> init([]() {
@@ -496,8 +460,11 @@ public:
         std::forward<Func>(f),
         nanobind::name(name), // nanobind::scope(thisClass),
         extra...);
-    nanobind::object builtins = nanobind::module_::import_("builtins");
-    thisClass.attr(name) = builtins.attr("classmethod")(cf);
+    static SafeInit<nanobind::object> classmethodFn([]() {
+      return std::make_unique<nanobind::object>(
+          nanobind::module_::import_("builtins").attr("classmethod"));
+    });
+    thisClass.attr(name) = classmethodFn.get()(cf);
     return *this;
   }
